feature/refactor-easy-swoole
- Refactor EasySwoole
